move simplex_volume function to triangulation file
